Drop the "index" mention on INVALID_ID
We can use it both with start (id) and trigger (file) on our end.
The reply from KFMon is always INVALID_ID, because ultimately a filename
lookup points to an index, but internal log messages will make the
distinction properly.

TL;DR: Keep it vague here, when in doubt, check KFMon's log.